Add schema attributes file
